Box can HTTP POST the firmware to the web

+++ entrypoint.sh
@@ -10,12 +10,18 @@ git clone https://github.com/nemo-k-org/firmware.git
 cd firmware
 make build
 
-rm -f /artefacts/firmware.bin /artefacts/firmware.bin.sha256 /artefacts/nemo-k-firmware.zip
-cp .pio/build/wemosd1/firmware.bin /artefacts/
-
-cd /artefacts
+mkdir temp-artefacts/
+cp .pio/build/wemosd1/firmware.bin temp-artefacts/
+cd temp-artefacts/
 sha256sum firmware.bin >firmware.bin.sha256
 zip nemo-k-firmware firmware.bin firmware.bin.sha256
+cd ..
 
-ARTEFACTS_OWNER=`stat --format=%u:%g .`
-chown $ARTEFACTS_OWNER firmware.bin firmware.bin.sha256 nemo-k-firmware.zip
+if [ -z $NEMOK_UPLOAD_URL ]; then
+    rm -f /artefacts/firmware.bin /artefacts/firmware.bin.sha256 /artefacts/nemo-k-firmware.zip
+    cp temp-artefacts/* /artefacts/
+    ARTEFACTS_OWNER=`stat --format=%u:%g /artefacts/`
+    chown $ARTEFACTS_OWNER /artefacts/firmware.bin /artefacts/firmware.bin.sha256 /artefacts/nemo-k-firmware.zip
+else
+    curl -X POST -F "firmware=@temp-artefacts/nemo-k-firmware.zip" "$NEMOK_UPLOAD_URL"
+fi
